+++ src/blocks.h
@@ -5,4 +5,4 @@
 
 pybind11::array_t<float> blockwise_means(const pybind11::array_t<int>& blocks,
                                          const pybind11::array_t<float>& input);
-pybind11::array_t<int> blocks_2d(const pybind11::array_t<float>& matrix);
+pybind11::array_t<int> blocks_2d(const pybind11::array_t<double>& matrix);
